initial migration test

+++ src/erc721/interfaces/ITRIS.sol
@@ -7,7 +7,7 @@ import {IERC721} from "openzeppelin-contracts/contracts/token/ERC721/IERC721.sol
 interface ITRIS is IERC721 {
     function owner() external returns (address);
 
-    function adminMint(address, uint) external;
+    function adminMint(address, uint256) external;
 
-    function mint(address, uint) external;
+    function mint(address, uint256) external;
 }
